+++ hometogo-frontend/src/pages/AdminPortal.tsx
@@ -1,16 +1,37 @@
+/**
+ * useful imports that Jared explained how to use 
+ * combined with imports that we suggested by online tutorials
+ * 
+ * useState useEffect --> tsx is stupid cannot remmeber things unless you
+ * tell it to. 
+ * axios --> talking to data, in this case mySQL database
+ */
 import React, { useState, useEffect } from "react";
 import axios from "axios";
 import { Button } from "@/components/ui/button";
 import {
   Dialog,
   DialogContent,
-  DialogDescription,
   DialogHeader,
   DialogTitle,
   DialogTrigger,
 } from "@/components/ui/dialog";
 import { Input } from "@/components/ui/input";
 
+
+
+/**
+ * Obejcts! 
+ * Similar to structs in C++ 
+ * I defined this with the attributes of user deatils I am 
+ * attempting to pull from the database. 
+ * Admin realisticall should be able to view all users, properties, 
+ * and bookings made on those properties. 
+ * 
+ * IMPORTANT NOTE: 
+ * I have only been able to delete and edit a user!
+ * Will reuse the delete to ban users until banning is implemented.
+ */
 type User = {
   id: number;
   firstName: string;
@@ -19,31 +40,25 @@ type User = {
   username: string;
 };
 
-type Host = {
-  hostID: number;
-  user: User;
-};
+type Host = { hostID: number; user: User };
+type Tenant = { tenantID: number; user: User };
+type Property = { id: number; name: string; location: string };
+type Booking = { id: number; date: string; status: string };
 
-type Tenant = {
-  tenantID: number;
-  user: User;
-};
 
-type Property = {
-  id: number;
-  name: string;
-  location: string;
-};
 
-type Booking = {
-  id: number;
-  date: string;
-  status: string;
-};
 
+
+//ADMIN PORTAL
 const AdminPortal: React.FC = () => {
+
+
+
+    //Literal tabs
   const [tab, setTab] = useState<"users" | "hosts" | "tenants" | "properties" | "bookings">("users");
 
+
+  // data states for the enties that Admin is to modrate 
   const [users, setUsers] = useState<User[]>([]);
   const [hosts, setHosts] = useState<Host[]>([]);
   const [tenants, setTenants] = useState<Tenant[]>([]);
@@ -51,59 +66,104 @@ const AdminPortal: React.FC = () => {
   const [bookings, setBookings] = useState<Booking[]>([]);
   const [selectedPropertyId, setSelectedPropertyId] = useState<number | null>(null);
 
-  // Fetchers
+
+
+
+  // Data states intened to keeping stats and other anlytics, meant for show, required by teacher guidlines
+  const [stats, setStats] = useState({
+    users: 0,
+    hosts: 0,
+    tenants: 0,
+    properties: 0,
+    bookings: 0,
+  });
+
+
+
+
+
+  // Meant to fetch stats and gather them
+  const fetchStats = async () => {
+    const [u, h, t, p] = await Promise.all([
+      axios.get("http://localhost:8080/api/users"),
+      axios.get("http://localhost:8080/api/host"),
+      axios.get("http://localhost:8080/api/tenant"),
+      axios.get("http://localhost:8080/api/properties"),
+    ]);
+    setStats({
+      users: u.data.length,
+      hosts: h.data.length,
+      tenants: t.data.length,
+      properties: p.data.length,
+      bookings: 0, // update if you add a count endpoint
+    });
+  };
+
+
+  // Talking to the database via our predeined endpoints from Java Crud API
   const fetchUsers = async () => {
     const res = await axios.get("http://localhost:8080/api/users");
     setUsers(res.data);
   };
-
   const fetchHosts = async () => {
     const res = await axios.get("http://localhost:8080/api/host");
     setHosts(res.data);
   };
-
   const fetchTenants = async () => {
     const res = await axios.get("http://localhost:8080/api/tenant");
     setTenants(res.data);
   };
-
   const fetchProperties = async () => {
     const res = await axios.get("http://localhost:8080/api/properties");
     setProperties(res.data);
   };
-
   const fetchBookings = async (propertyId: number) => {
     const res = await axios.get(`http://localhost:8080/api/bookings/${propertyId}`);
     setBookings(res.data);
   };
 
-  // Delete user
   const deleteUser = async (id: number) => {
     await axios.delete(`http://localhost:8080/api/users/${id}`);
     fetchUsers();
+    fetchStats();
   };
 
+  // Constant call to fethStats
   useEffect(() => {
+    fetchStats();
     switch (tab) {
-      case "users":
-        fetchUsers();
-        break;
-      case "hosts":
-        fetchHosts();
-        break;
-      case "tenants":
-        fetchTenants();
-        break;
-      case "properties":
-        fetchProperties();
-        break;
+      case "users": fetchUsers(); break;
+      case "hosts": fetchHosts(); break;
+      case "tenants": fetchTenants(); break;
+      case "properties": fetchProperties(); break;
     }
   }, [tab]);
 
+
+
+
+  // Render the UI for the admin page
+  /**
+   * Set of tables that will be populated with the 
+   * data from the database.
+   */
   return (
     <div className="p-6 max-w-7xl mx-auto">
       <h1 className="text-3xl font-bold mb-6 text-center">Admin Dashboard</h1>
 
+      {/* Stats */}
+      <div className="grid grid-cols-2 sm:grid-cols-3 md:grid-cols-5 gap-4 mb-6">
+        {Object.entries(stats).map(([key, val]) => (
+          <div key={key} className="bg-gray-100 rounded-lg p-4 text-center">
+            <h3 className="text-xl font-bold">{val}</h3>
+            <p className="text-muted-foreground capitalize">{key}</p>
+          </div>
+        ))}
+      </div>
+
+
+        
+      {/* Tabs */}
       <div className="flex justify-center mb-4 gap-2 flex-wrap">
         {["users", "hosts", "tenants", "properties", "bookings"].map((item) => (
           <Button
@@ -116,52 +176,67 @@ const AdminPortal: React.FC = () => {
         ))}
       </div>
 
-      {/* USERS */}
+      {/* Users */}
       {tab === "users" && (
-        <>
-          <div className="flex justify-end mb-3">
-            <Dialog>
-              <DialogTrigger asChild>
-                <Button>Add User</Button>
-              </DialogTrigger>
-              <DialogContent>
-                <DialogHeader>
-                  <DialogTitle>Add User</DialogTitle>
-                  <DialogDescription>Form pending implementation</DialogDescription>
-                </DialogHeader>
-              </DialogContent>
-            </Dialog>
-          </div>
-          <table className="w-full border">
-            <thead className="bg-gray-100">
-              <tr>
-                <th className="p-2 text-left">ID</th>
-                <th className="p-2 text-left">Name</th>
-                <th className="p-2 text-left">Email</th>
-                <th className="p-2 text-left">Actions</th>
-              </tr>
-            </thead>
-            <tbody>
-              {users.map((u) => (
-                <tr key={u.id} className="border-t">
-                  <td className="p-2">{u.id}</td>
-                  <td className="p-2">{u.firstName} {u.lastName}</td>
-                  <td className="p-2">{u.email}</td>
-                  <td className="p-2">
-                    <Button variant="destructive" size="sm" onClick={() => deleteUser(u.id)}>
+        <table className="w-full border text-sm">
+          <thead className="bg-gray-100">
+            <tr>
+              <th className="p-2 text-left">ID</th>
+              <th className="p-2 text-left">Name</th>
+              <th className="p-2 text-left">Email</th>
+              <th className="p-2 text-left">Username</th>
+              <th className="p-2 text-left">Actions</th>
+            </tr>
+          </thead>
+          <tbody>
+            {users.map((user) => {
+              const updatedUser = { ...user };
+              return (
+                <tr key={user.id} className="border-t">
+                  <td className="p-2">{user.id}</td>
+                  <td className="p-2">{user.firstName} {user.lastName}</td>
+                  <td className="p-2">{user.email}</td>
+                  <td className="p-2">{user.username}</td>
+                  <td className="p-2 space-x-2">
+                    <Dialog>
+                      <DialogTrigger asChild>
+                        <Button variant="outline" size="sm">Edit</Button>
+                      </DialogTrigger>
+                      <DialogContent>
+                        <DialogHeader>
+                          <DialogTitle>Edit User</DialogTitle>
+                        </DialogHeader>
+                        <form
+                          className="space-y-3"
+                          onSubmit={async (e) => {
+                            e.preventDefault();
+                            await axios.put(`http://localhost:8080/api/users/${user.id}`, updatedUser);
+                            fetchUsers();
+                            fetchStats();
+                          }}
+                        >
+                          <Input defaultValue={user.firstName} onChange={(e) => updatedUser.firstName = e.target.value} />
+                          <Input defaultValue={user.lastName} onChange={(e) => updatedUser.lastName = e.target.value} />
+                          <Input defaultValue={user.email} onChange={(e) => updatedUser.email = e.target.value} />
+                          <Input defaultValue={user.username} onChange={(e) => updatedUser.username = e.target.value} />
+                          <Button type="submit">Save</Button>
+                        </form>
+                      </DialogContent>
+                    </Dialog>
+                    <Button variant="destructive" size="sm" onClick={() => deleteUser(user.id)}>
                       Delete
                     </Button>
                   </td>
                 </tr>
-              ))}
-            </tbody>
-          </table>
-        </>
+              );
+            })}
+          </tbody>
+        </table>
       )}
 
-      {/* HOSTS */}
+      {/* Hosts */}
       {tab === "hosts" && (
-        <table className="w-full border">
+        <table className="w-full border text-sm">
           <thead className="bg-gray-100">
             <tr>
               <th className="p-2 text-left">Host ID</th>
@@ -181,9 +256,9 @@ const AdminPortal: React.FC = () => {
         </table>
       )}
 
-      {/* TENANTS */}
+      {/* Tenants */}
       {tab === "tenants" && (
-        <table className="w-full border">
+        <table className="w-full border text-sm">
           <thead className="bg-gray-100">
             <tr>
               <th className="p-2 text-left">Tenant ID</th>
@@ -203,15 +278,15 @@ const AdminPortal: React.FC = () => {
         </table>
       )}
 
-      {/* PROPERTIES */}
+      {/* Properties */}
       {tab === "properties" && (
-        <table className="w-full border">
+        <table className="w-full border text-sm">
           <thead className="bg-gray-100">
             <tr>
               <th className="p-2 text-left">Property ID</th>
               <th className="p-2 text-left">Name</th>
               <th className="p-2 text-left">Location</th>
-              <th className="p-2 text-left">View Bookings</th>
+              <th className="p-2 text-left">Bookings</th>
             </tr>
           </thead>
           <tbody>
@@ -233,11 +308,11 @@ const AdminPortal: React.FC = () => {
         </table>
       )}
 
-      {/* BOOKINGS */}
+      {/* Bookings */}
       {tab === "bookings" && selectedPropertyId && (
         <>
           <h2 className="text-xl font-semibold mb-4">Bookings for Property #{selectedPropertyId}</h2>
-          <table className="w-full border">
+          <table className="w-full border text-sm">
             <thead className="bg-gray-100">
               <tr>
                 <th className="p-2 text-left">Booking ID</th>
